refactor: add sub directory entity and move entities into it
